requirements.txt
autopep8==1.7.0
bottle==0.12.23
pycodestyle==2.9.1
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
